Adding authentication with JWT

+++ app/Http/Controllers/MeetingController.php
@@ -5,12 +5,15 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
 use App\Meeting;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
 
     public function __construct(){
-
+        $this->middleware('jwt.auth', ['only' => [
+            "store", "update", "destroy"
+        ]]);
     }
     /**
      * Display a listing of the resource.
@@ -49,12 +52,16 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required',
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'],404);
+        }
+
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'title' => $title,
@@ -116,13 +123,17 @@ class MeetingController extends Controller
         $this->validate($request,[
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required',
+            'time' => 'required|date_format:YmdHie',            
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'],404);
+        }
+
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -159,6 +170,15 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'],404);
+        }
+
+        if(!$meeting->users()->where('users.id',$user->id)->first()){
+            return response()->json(['msg' => 'User is not registered for this meeting. Update failed'], 401);
+        }
+
         $users = $meeting->users;
         $meeting->users->detach();
         if(!$meeting->delete()){
